pyproject.toml
[project]
name = "gif-to-webp-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow==12.0.0",
]
